Refactoring, connect ResultView to PodcastDescriptionView.
